requirements.txt
flask
openai
numpy
python-dotenv
gunicorn